+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2024-03-31
+
+### Added
+- 🏗️ **Amélioration des types Schema.org**
+  - Ajout de 25+ sous-types spécialisés
+  - Organisation par catégories avec optgroups
+  - Architecture & Bâtiment (ArchitecturalService, GeneralContractor, etc.)
+  - Services Professionnels (LegalService, AccountingService, etc.)
+  - Santé & Bien-être (Dentist, BeautySalon, FitnessCenter, etc.)
+  - Commerce & Restauration (Bakery, CafeOrCoffeeShop, Hotel, etc.)
+  - Éducation & Loisirs (EducationalOrganization, ArtGallery, etc.)
+
 ## [1.0.1] - 2024-03-31
 
 ### Added